+++ materialui/mui-popover.lua
@@ -1,7 +1,7 @@
 --[[
     A loosely based Material UI module
 
-    mui-select.lua : This is for creating select drop-downs (pick lists)
+    mui-popover.lua : This is for creating popover menus (pick lists)
 
     The MIT License (MIT)
 
@@ -42,11 +42,11 @@ local mathABS = math.abs
 
 local M = muiData.M -- {} -- for module array/table
 
-function M.createSelect(options)
-    M.newSelect(options)
+function M.createPopover(options)
+    M.newPopOver(options)
 end
 
-function M.newSelect(options)
+function M.newPopover(options)
 
     local x,y = 160, 240
     if options.x ~= nil then
@@ -70,7 +70,7 @@ function M.newSelect(options)
 
     muiData.widgetDict[options.name] = {}
     muiData.widgetDict[options.name].name = options.name
-    muiData.widgetDict[options.name]["type"] = "Selector"
+    muiData.widgetDict[options.name]["type"] = "Popover"
 
     muiData.widgetDict[options.name]["container"] = display.newContainer(options.width+4, options.height + options.listHeight)
     muiData.widgetDict[options.name]["container"]:translate( x, y ) -- center the container
@@ -102,106 +102,34 @@ function M.newSelect(options)
         options.strokeColor = { 0.4, 0.4, 0.4, 1 }
     end
 
-    muiData.widgetDict[options.name]["rect"] = display.newRect( 0, 0, options.width, options.height )
-    muiData.widgetDict[options.name]["rect"]:setFillColor( unpack( options.fieldBackgroundColor ) )
-    muiData.widgetDict[options.name]["container"]:insert( muiData.widgetDict[options.name]["rect"] )
+    options.rowColor = options.backgroundColor
 
-    local rect = muiData.widgetDict[options.name]["rect"]
-    muiData.widgetDict[options.name]["line"] = display.newLine( -(rect.contentWidth * 0.9), rect.contentHeight / 2, (rect.contentWidth * 0.5), rect.contentHeight / 2)
-    muiData.widgetDict[options.name]["line"].strokeWidth = M.getScaleVal(4)
-    muiData.widgetDict[options.name]["line"]:setStrokeColor( unpack(options.inactiveColor) )
-    muiData.widgetDict[options.name]["container"]:insert( muiData.widgetDict[options.name]["line"] )
-
-    local labelOptions =
-    {
-        --parent = textGroup,
-        text = options.labelText,
-        x = -(rect.contentWidth * 0.25),
-        y = -(rect.contentHeight * 0.95),
-        width = rect.contentWidth * 0.5,     --required for multi-line and alignment
-        font = options.font,
-        fontSize = options.height * 0.55,
-        align = "left"  --new alignment parameter
-    }
-    muiData.widgetDict[options.name]["textlabel"] = display.newText( labelOptions )
-    muiData.widgetDict[options.name]["textlabel"]:setFillColor( unpack(options.inactiveColor) )
-    muiData.widgetDict[options.name]["textlabel"].inactiveColor = options.inactiveColor
-    muiData.widgetDict[options.name]["textlabel"].activeColor = options.activeColor
-    muiData.widgetDict[options.name]["container"]:insert( muiData.widgetDict[options.name]["textlabel"] )
-
-    local scaleFontSize = 1
-    if muiData.environment == "simulator" then
-        scaleFontSize = 0.75
-    end
-
-    local textOptions =
-    {
-        --parent = textGroup,
-        text = options.text,
-        x = 0,
-        y = 0,
-        width = options.width,
-        font = options.font,
-        fontSize = options.height * 0.55,
-        align = "left"  --new alignment parameter
-    }
-    muiData.widgetDict[options.name]["selectorfieldfake"] = display.newText( textOptions )
-    muiData.widgetDict[options.name]["selectorfieldfake"]:setFillColor( unpack(muiData.widgetDict[options.name]["textlabel"].inactiveColor) )
-    muiData.widgetDict[options.name]["selectorfieldfake"]:addEventListener("touch", M.selectorListener)
-    muiData.widgetDict[options.name]["selectorfieldfake"].name = options.name
-    muiData.widgetDict[options.name]["selectorfieldfake"].dialogName = options.dialogName
-    muiData.widgetDict[options.name]["container"]:insert( muiData.widgetDict[options.name]["selectorfieldfake"] )
-
-    -- use codepoints like keyboard_arrow_down
-    textOptions =
-    {
-        --parent = textGroup,
-        text = "keyboard_arrow_down",
-        x = 0,
-        y = 0,
-        width = options.width,
-        font = "MaterialIcons-Regular.ttf",
-        fontSize = options.height * 0.55,
-        align = "right"  --new alignment parameter
-    }
-    muiData.widgetDict[options.name]["selectorfieldarrow"] = display.newText( textOptions )
-    muiData.widgetDict[options.name]["selectorfieldarrow"]:setFillColor( unpack(muiData.widgetDict[options.name]["textlabel"].inactiveColor) )
-    muiData.widgetDict[options.name]["selectorfieldarrow"].name = options.name
-    muiData.widgetDict[options.name]["selectorfieldarrow"].dialogName = options.dialogName
-    muiData.widgetDict[options.name]["container"]:insert( muiData.widgetDict[options.name]["selectorfieldarrow"] )
-
-    if options.listHeight > muiData.contentHeight then
-        options.listHeight = muiData.contentHeight * 0.75
-    end
-
+    muiData.dialogInUse = true
+    muiData.dialogName = options.name
     muiData.widgetDict[options.name]["options"] = options
+    M.revealTableViewForPopover(options.name, muiData.widgetDict[options.name]["options"])
 end
 
-function M.getSelectorProperty(widgetName, propertyName)
+function M.getPopoverProperty(widgetName, propertyName)
     local data = nil
 
     if widgetName == nil or propertyName == nil then return data end
 
     if propertyName == "object" then
         data = muiData.widgetDict[widgetName]["container"] -- x,y movement
-    elseif propertyName == "label" then
-        data = muiData.widgetDict[widgetName]["textlabel"] -- label
-    elseif propertyName == "value" then
-        data = muiData.widgetDict[widgetName]["value"] -- value
-    elseif propertyName == "layer_1" then
-        data = muiData.widgetDict[widgetName]["rect"] -- clickable area
-    elseif propertyName == "layer_2" then
-        data = muiData.widgetDict[widgetName]["selectorfieldarrow"] -- icon arrow
-    elseif propertyName == "layer_3" then
-        data = muiData.widgetDict[widgetName]["line"] -- line beneath control
-    elseif propertyName == "layer_4" then
-        data = muiData.widgetDict[widgetName]["mygroup"]
     end
+
     return data
 end
 
-function M.revealTableViewForSelector(name, options)
-    -- table view to hold pick list keyboard_arrow_down
+function M.revealTableViewForPopover(name, options)
+    muiData.widgetDict[options.name]["rectback"] = display.newRect( display.contentCenterX, display.contentCenterY, muiData.contentWidth, muiData.contentHeight)
+    muiData.widgetDict[options.name]["rectback"].strokeWidth = 0
+    muiData.widgetDict[options.name]["rectback"]:setStrokeColor( 1 )
+    muiData.widgetDict[options.name]["rectback"].alpha = 0.01
+    muiData.widgetDict[options.name]["rectback"].muiName = options.name
+    muiData.widgetDict[options.name]["rectback"]:addEventListener( "touch", M.touchPopoverBack )
+
     muiData.widgetDict[options.name]["mygroup"] = display.newGroup() -- options.width+4, options.height + options.listHeight)
 
     local x = options.x
@@ -211,25 +139,28 @@ function M.revealTableViewForSelector(name, options)
         y = muiData.widgetDict[options.name]["calculated"].y
     end
 
+    options.leftMargin = options.leftMargin or M.getScaleVal(20)
+
     muiData.widgetDict[options.name]["mygroup"].x = x
     muiData.widgetDict[options.name]["mygroup"].y = y
 
     M.newTableView({
         name = options.name.."-List",
-        width = options.width - M.getScaleVal(5),
+        width = options.width - (options.leftMargin * 2),
         height = options.listHeight,
         font = options.font,
-        top = M.getScaleVal(40),
-        left = 0,
+        top = M.getScaleVal(30),
+        left = options.leftMargin,
         textColor = options.textColor,
         strokeColor = options.inactiveColor,
         strokeWidth = 1,
         lineHeight = 0,
         noLines = true,
+        isLocked = true,
         rowColor = options.rowColor,
         rowHeight = options.height,
         callBackTouch = options.callBackTouch,
-        callBackRender = M.onRowRenderSelect,
+        callBackRender = M.onRowRenderPopover,
         scrollListener = options.listener,
         categoryColor = options.categoryColor,
         categoryLineColor = options.categoryLineColor,
@@ -240,6 +171,7 @@ function M.revealTableViewForSelector(name, options)
     muiData.widgetDict[options.name]["rect2"] = display.newRect( options.width * 0.5, (options.listHeight * 0.45) + options.height, options.width, options.listHeight + (options.height * 0.5))
     muiData.widgetDict[options.name]["rect2"].strokeWidth = options.strokeWidth
     muiData.widgetDict[options.name]["rect2"]:setStrokeColor( unpack( options.strokeColor ) )
+    muiData.widgetDict[options.name]["rect2"]:setFillColor( unpack( options.backgroundColor ) )
     muiData.widgetDict[options.name]["mygroup"]:insert( muiData.widgetDict[options.name]["rect2"] )
 
     if muiData.widgetDict[options.name]["calculated"] == nil then
@@ -265,19 +197,21 @@ function M.revealTableViewForSelector(name, options)
         muiData.widgetDict[options.name]["calculated"].y = muiData.widgetDict[options.name]["mygroup"].y
     end
     muiData.widgetDict[options.name]["mygroup"]:insert( muiData.widgetDict[options.name.."-List"]["tableview"] )
-    muiData.widgetDict[options.name]["mygroup"].isVisible = false
 end
 
-function M.selectorListener( event )
-    if event.phase == "began" then
-        local name = event.target.name
-        muiData.currentTargetName = name
-        M.revealTableViewForSelector(name, muiData.widgetDict[name]["options"])
-        muiData.widgetDict[name]["mygroup"].isVisible = true
+function M.touchPopoverBack( event )
+    if event.target.muiName ~= nil then
+
+        if ( event.phase == "ended" ) then
+            M.removePopover(event.target.muiName)
+            muiData.dialogInUse = false
+            muiData.dialogName = nil
+            return true -- prevent propagation to other controls
+        end
     end
 end
 
-function M.onRowRenderSelect( event )
+function M.onRowRenderPopover( event )
     local row = event.row
 
     -- Cache the row "contentWidth" and "contentHeight" because the row bounds can change as children objects are added
@@ -294,7 +228,7 @@ function M.onRowRenderSelect( event )
     rowTitle.y = rowHeight * 0.5
 end
 
-function M.onRowTouchSelector(event)
+function M.onRowTouchPopover(event)
     local muiTarget = M.getEventParameter(event, "muiTarget")
     local muiTargetValue = M.getEventParameter(event, "muiTargetValue")
     local muiTargetIndex = M.getEventParameter(event, "muiTargetIndex")
@@ -309,56 +243,47 @@ function M.onRowTouchSelector(event)
 
     if event.row.miscEvent ~= nil and event.row.miscEvent.name ~= nil then
         local parentName = string.gsub(event.row.miscEvent.name, "-List", "")
-
-        muiData.widgetDict[parentName]["selectorfieldfake"].text = muiTargetValue
-        muiData.widgetDict[parentName]["value"] = muiTargetValue
-        timer.performWithDelay(500, function() M.finishSelector(parentName) end, 1)
+        timer.performWithDelay(500, function() M.finishPopover(parentName) end, 1)
     end
+    muiData.dialogInUse = false
+    muiData.dialogName = nil
+    return true -- prevent propagation to other controls
 end
 
-function M.finishSelector(parentName)
+function M.finishPopover(parentName)
     if muiData.widgetDict[parentName] == nil then return end
-    M.removeSelector(parentName, "listonly")
+    M.removePopover(parentName, "listonly")
 end
 
-function M.removeWidgetSelector(widgetName, listonly)
-    M.removeSelector(widgetName, listonly)
+function M.removeWidgetPopover(widgetName, listonly)
+    M.removePopover(widgetName, listonly)
 end
 
-function M.removeSelector(widgetName, listonly)
+function M.removePopover(widgetName, listonly)
+
     if widgetName == nil then
         return
     end
 
     if muiData.widgetDict[widgetName] == nil then return end
 
+    muiData.widgetDict[widgetName]["menuactive"] = false
+
     if listonly ~= nil then
         M.removeTableView(widgetName .. "-List")
-        M.removeSelectorGroup(widgetName)
+        M.removePopoverGroup(widgetName)
         return
     else
         M.removeTableView(widgetName .. "-List")
     end
 
-    muiData.widgetDict[widgetName]["selectorfieldfake"]:removeEventListener("touch", M.selectorListener)
-
-    muiData.widgetDict[widgetName]["selectorfieldarrow"]:removeSelf()
-    muiData.widgetDict[widgetName]["selectorfieldarrow"] = nil
-    muiData.widgetDict[widgetName]["selectorfieldfake"]:removeSelf()
-    muiData.widgetDict[widgetName]["selectorfieldfake"] = nil
-    muiData.widgetDict[widgetName]["textlabel"]:removeSelf()
-    muiData.widgetDict[widgetName]["textlabel"] = nil
-    muiData.widgetDict[widgetName]["rect"]:removeSelf()
-    muiData.widgetDict[widgetName]["rect"] = nil
-    muiData.widgetDict[widgetName]["line"]:removeSelf()
-    muiData.widgetDict[widgetName]["line"] = nil
-    M.removeSelectorGroup(widgetName)
+    M.removePopoverGroup(widgetName)
     muiData.widgetDict[widgetName]["container"]:removeSelf()
     muiData.widgetDict[widgetName]["container"] = nil
     muiData.widgetDict[widgetName] = nil
 end
 
-function M.removeSelectorGroup(widgetName)
+function M.removePopoverGroup(widgetName)
     if widgetName == nil then
         return
     end
@@ -369,6 +294,13 @@ function M.removeSelectorGroup(widgetName)
         muiData.widgetDict[widgetName]["rect2"]:removeSelf()
         muiData.widgetDict[widgetName]["rect2"] = nil
     end
+
+    muiData.widgetDict[widgetName]["rectback"]:removeEventListener("touch", M.touchPopoverBack)
+    if muiData.widgetDict[widgetName]["rectback"] ~= nil then
+        muiData.widgetDict[widgetName]["rectback"]:removeSelf()
+        muiData.widgetDict[widgetName]["rectback"] = nil
+    end
+
     if muiData.widgetDict[widgetName]["mygroup"] ~= nil then
         muiData.widgetDict[widgetName]["mygroup"]:removeSelf()
         muiData.widgetDict[widgetName]["mygroup"] = nil